Completely refactored the StaticApi.
- The StaticApi is now a  property on the RiotApi (api.Static)
- All Endpoints of the  StaticApi are now properties of the IStaticDataEndpoint (api.Static.Champion, ...)
- The StaticDataEndpoint fully supports dependency injection.
- Every static Endpoint has its own configuration (SlidingExpirationTime, Cache, Requester)
- All static Endpoints derive from IStaticEndpoint (empty interface)
