+++ RiotSharp.Test/StaticRiotApiExceptionTest.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using RiotSharp.Endpoints.StaticDataEndpoint;
 using RiotSharp.Misc;
 
 namespace RiotSharp.Test
@@ -6,14 +7,14 @@ namespace RiotSharp.Test
     [TestClass]
     public class StaticRiotApiExceptionTest
     {
-        private static readonly StaticRiotApi FaultyStaticApi = StaticRiotApi.GetInstance(CommonTestBase.FaultyApiKey);
+        private static readonly StaticDataEndpoint FaultyStaticApi = StaticDataEndpoint.GetInstance(CommonTestBase.FaultyApiKey);
 
         [TestMethod]
         [TestCategory("Exception")]
         [ExpectedException(typeof(RiotSharpException))]
         public void GetStatic_ShouldThrowRiotSharpException_Test()
         {
-            FaultyStaticApi.GetChampionAsync(Region.euw, 1).GetAwaiter().GetResult();
+            FaultyStaticApi.Champion.GetChampionAsync(Region.euw, 1).GetAwaiter().GetResult();
         }
     }
 }
