+++ RiotSharp/Endpoints/Interfaces/Static/IStaticEndpoint.cs
@@ -1,0 +1,6 @@
+﻿namespace RiotSharp.Endpoints.Interfaces.Static
+{
+    public interface IStaticEndpoint
+    {
+    }
+}
