+++ RiotSharp/Endpoints/Interfaces/Static/IStaticSummonerSpellEndpoint.cs
@@ -1,0 +1,31 @@
+﻿using RiotSharp.Endpoints.StaticDataEndpoint;
+using RiotSharp.Endpoints.StaticDataEndpoint.SummonerSpell;
+using RiotSharp.Misc;
+using System.Threading.Tasks;
+
+namespace RiotSharp.Endpoints.Interfaces.Static
+{
+    public interface IStaticSummonerSpellEndpoint : IStaticEndpoint
+    {
+        /// <summary>
+        /// Get a summoner spell asynchronously.
+        /// </summary>
+        /// <param name="region">Region from which to retrieve the data.</param>
+        /// <param name="summonerSpell">Summoner spell to retrieve.</param>
+        /// <param name="summonerSpellData">Data to retrieve.</param>
+        /// <param name="language">Language of the data to be retrieved.</param>
+        /// <returns>A summoner spell.</returns>
+        Task<SummonerSpellStatic> GetSummonerSpellAsync(Region region, int summonerSpellId,
+            SummonerSpellData summonerSpellData = SummonerSpellData.All, Language language = Language.en_US);
+
+        /// <summary>
+        /// Get a list of all summoner spells asynchronously.
+        /// </summary>
+        /// <param name="region">Region from which to retrieve the data.</param>
+        /// <param name="summonerSpellData">Data to retrieve.</param>
+        /// <param name="language">Language of the data to be retrieved.</param>
+        /// <returns>A SummonerSpellListStatic object containing all summoner spells.</returns>
+        Task<SummonerSpellListStatic> GetSummonerSpellsAsync(Region region,
+            SummonerSpellData summonerSpellData = SummonerSpellData.All, Language language = Language.en_US);
+    }
+}
