+++ RiotSharp/Endpoints/Interfaces/Static/IStaticVersionEndpoint.cs
@@ -1,0 +1,11 @@
+﻿using RiotSharp.Misc;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace RiotSharp.Endpoints.Interfaces.Static
+{
+    public interface IStaticVersionEndpoint : IStaticEndpoint
+    {
+        Task<List<string>> GetVersionsAsync(Region region);
+    }
+}
